+++ res/animation/player.json
