+++ res/animation/snibee.json
