+++ src/graphics/animationParser.ts
@@ -9,7 +9,7 @@ type SpriteSetting = {
   default: string
 }
 
-export function parseSprite(json: SpriteSetting): Animation {
+export function parseAnimation(json: SpriteSetting): Animation {
   checkMembers(json, { name: 'string', state: 'any', default: 'string' }, 'sprite')
 
   const name = json.name
